move SeqNoPrimaryTermAndIndex into indexer
