+++ x-pack/plugin/transform/src/test/java/org/elasticsearch/xpack/transform/transforms/ClientDataFrameIndexerTests.java
@@ -67,6 +67,7 @@ public class ClientDataFrameIndexerTests extends ESTestCase {
                 2L,
                 Collections.emptyMap(),
                 Instant.now().toEpochMilli()),
+            null,
             parentTask);
 
         List<Boolean> shouldAudit = IntStream.range(0, 100_000).boxed().map(indexer::shouldAuditOnFinish).collect(Collectors.toList());
